+++ src/geometry/quaternion_construction.rs
@@ -351,7 +351,7 @@ impl<N: Real> UnitQuaternion<N> {
 
     /// Creates a new unit quaternion rotation from a rotation axis scaled by the rotation angle.
     ///
-    /// If `axisangle` is zero, this returns the indentity rotation.
+    /// If `axisangle` has a magnitude smaller than `N::default_epsilon()`, this returns the indentity rotation.
     #[inline]
     pub fn new<SB>(axisangle: Vector<N, U3, SB>) -> Self
     where
@@ -364,7 +364,20 @@ impl<N: Real> UnitQuaternion<N> {
 
     /// Creates a new unit quaternion rotation from a rotation axis scaled by the rotation angle.
     ///
-    /// If `axisangle` is zero, this returns the indentity rotation.
+    /// If `axisangle` has a magnitude smaller than `eps`, this returns the indentity rotation.
+    #[inline]
+    pub fn new_eps<SB>(axisangle: Vector<N, U3, SB>, eps: N) -> Self
+    where
+        SB: Storage<N, U3>,
+    {
+        let two: N = ::convert(2.0f64);
+        let q = Quaternion::<N>::from_parts(N::zero(), axisangle / two).exp_eps(eps);
+        Self::new_unchecked(q)
+    }
+
+    /// Creates a new unit quaternion rotation from a rotation axis scaled by the rotation angle.
+    ///
+    /// If `axisangle` has a magnitude smalle than `N::default_epsilon()`, this returns the indentity rotation.
     /// Same as `Self::new(axisangle)`.
     #[inline]
     pub fn from_scaled_axis<SB>(axisangle: Vector<N, U3, SB>) -> Self
@@ -372,6 +385,18 @@ impl<N: Real> UnitQuaternion<N> {
         SB: Storage<N, U3>,
     {
         Self::new(axisangle)
+    }
+
+    /// Creates a new unit quaternion rotation from a rotation axis scaled by the rotation angle.
+    ///
+    /// If `axisangle` has a mangnitude smaller than `eps`, this returns the indentity rotation.
+    /// Same as `Self::new(axisangle)`.
+    #[inline]
+    pub fn from_scaled_axis_eps<SB>(axisangle: Vector<N, U3, SB>, eps: N) -> Self
+    where
+        SB: Storage<N, U3>,
+    {
+        Self::new_eps(axisangle, eps)
     }
 }